Module depy_gui::gui::elements::controller

Structs

AppController

Enums

ControllerError

Constants

ADD_BUCKET 🔒

ADD_PACKAGE 🔒

CLEAN_DEPY 🔒

FAILED_INSTALL 🔒

FAILED_SEARCH 🔒

FINISHED_CLEAN 🔒

FINISHED_INITIALIZE 🔒

FINISHED_INSTALL 🔒

FINISHED_SEARCH

INITIALIZE 🔒

REMOVE_BUCKET 🔒

REMOVE_PACKAGE 🔒

UNINSTALL_DEPY 🔒

UPDATE_BUCKETS 🔒

Functions

add_bucket 🔒

clean_depy_async 🔒

find_packages_async

flatten_err 🔒

Intended to be used to flatten the output of catch_unwind. The wrapped result will be proccesd and flattened down to a single Result. Uses the closure to proccess the unwind error into a flat error, the argument of the closure being the error message of the unwind error transformed to a string

init_depy_async 🔒

install_packages

remove_bucket 🔒

uninstall_depy_async 🔒