Module depy_gui::gui::theme

Constants

BACKGROUND_DARK

BACKGROUND_LIGHT

BASIC_WIDGET_HEIGHT

BORDERED_WIDGET_HEIGHT

BORDER_DARK

BORDER_LIGHT

BUTTON_BORDER_RADIUS

BUTTON_BORDER_WIDTH

BUTTON_DARK

BUTTON_LIGHT

CURSOR_COLOR

DISABLED_BUTTON_DARK

DISABLED_BUTTON_LIGHT

DISABLED_FOREGROUND_DARK

DISABLED_FOREGROUND_LIGHT

DISABLED_TEXT_COLOR

FOREGROUND_DARK

FOREGROUND_LIGHT

PLACEHOLDER_COLOR

PRIMARY_DARK

PRIMARY_LIGHT

PROGRESS_BAR_RADIUS

SCROLLBAR_BORDER_COLOR

SCROLLBAR_COLOR

SCROLLBAR_EDGE_WIDTH

SCROLLBAR_FADE_DELAY

SCROLLBAR_MAX_OPACITY

SCROLLBAR_MIN_SIZE

Minimum length for any scrollbar to be when measured on that scrollbar’s primary axis.

SCROLLBAR_PAD

SCROLLBAR_RADIUS

SCROLLBAR_WIDTH

SELECTED_TEXT_BACKGROUND_COLOR

SELECTED_TEXT_INACTIVE_BACKGROUND_COLOR

SELECTION_TEXT_COLOR

TEXTBOX_BORDER_RADIUS

TEXTBOX_BORDER_WIDTH

TEXTBOX_INSETS

TEXT_COLOR

TEXT_SIZE_LARGE

TEXT_SIZE_NORMAL

UI_FONT

The default font for labels, buttons, text boxes, and other UI elements.

UI_FONT_BOLD

A bold version of the default UI font.

UI_FONT_ITALIC

An Italic version of the default UI font.

WIDE_WIDGET_WIDTH

The default minimum width for a ‘wide’ widget; a textbox, slider, progress bar, etc.

WIDGET_CONTROL_COMPONENT_PADDING

The default internal (horizontal) padding for visually distinct components of a widget; for instance between a checkbox and its label.

WIDGET_PADDING_HORIZONTAL

The default horizontal spacing between widgets.

WIDGET_PADDING_VERTICAL

The default vertical spacing between widgets.

WINDOW_BACKGROUND_COLOR

Functions

setup_dark 🔒

Catpuccin Mocha

setup_light 🔒

Catpuccin Macchiato

setup_theme